+++ src/main/java/iwb/domain/db/W5TableField.java
@@ -41,7 +41,7 @@ public class W5TableField implements java.io.Serializable, W5Param, W5Base {
 
 	private Short minLength;
 
-	private Short maxLength;
+	private Integer maxLength;
 
 	private BigDecimal minValue;
 
@@ -145,11 +145,11 @@ public class W5TableField implements java.io.Serializable, W5Param, W5Base {
 	}
 
 	@Column(name="max_length")
-	public Short getMaxLength() {
+	public Integer getMaxLength() {
 		return this.maxLength;
 	}
 
-	public void setMaxLength(Short maxLength) {
+	public void setMaxLength(Integer maxLength) {
 		this.maxLength = maxLength;
 	}
